+++ src/app/public/modules/dropdown/fixtures/dropdown.component.fixture.ts
@@ -1,13 +1,22 @@
 import {
   ChangeDetectorRef,
   Component,
-  ElementRef,
-  ViewChild
+  QueryList,
+  ViewChild,
+  ViewChildren
 } from '@angular/core';
 
 import {
   Subject
 } from 'rxjs';
+
+import {
+  SkyPopoverAlignment
+} from '../../popover/types/popover-alignment';
+
+import {
+  SkyDropdownHorizontalAlignment
+} from '../types/dropdown-horizontal-alignment';
 
 import {
   SkyDropdownMessage
@@ -18,26 +27,71 @@ import {
 } from '../types/dropdown-message-type';
 
 import {
-  SkyDropdownComponent
-} from '../dropdown.component';
+  SkyDropdownTriggerType
+} from '../types/dropdown-trigger-type';
+
+import {
+  SkyDropdownItemComponent
+} from '../dropdown-item.component';
 
 import {
   SkyDropdownMenuComponent
 } from '../dropdown-menu.component';
 
+import {
+  SkyDropdownComponent
+} from '../dropdown.component';
+
 @Component({
   selector: 'sky-test-cmp',
   templateUrl: './dropdown.component.fixture.html'
 })
-export class DropdownTestComponent {
-  public alignment: string;
+export class DropdownFixtureComponent {
+
+  //#region directive properties
+
+  public alignment: SkyPopoverAlignment;
+
   public buttonStyle: string;
-  public buttonType: String;
+
+  public buttonType: string;
+
+  public disabled: boolean;
+
+  public dismissOnBlur: boolean;
+
+  public horizontalAlignment: SkyDropdownHorizontalAlignment;
+
+  public itemAriaRole: string;
+
   public label: string;
+
+  public messageStream = new Subject<SkyDropdownMessage>();
+
+  public menuAriaLabelledBy: string;
+
+  public menuAriaRole: string;
+
   public title: string;
-  public trigger: String;
-  public dropdownController = new Subject<SkyDropdownMessage>();
-  public isDisabled: boolean;
+
+  public trigger: SkyDropdownTriggerType;
+
+  public useNativeFocus: boolean;
+
+  //#endregion directive properties
+
+  @ViewChild('dropdownRef', {
+    read: SkyDropdownComponent
+  })
+  public dropdownRef: SkyDropdownComponent;
+
+  @ViewChild('dropdownMenuRef', {
+    read: SkyDropdownMenuComponent
+  })
+  public dropdownMenuRef: SkyDropdownMenuComponent;
+
+  @ViewChildren(SkyDropdownItemComponent)
+  public dropdownItemRefs: QueryList<SkyDropdownItemComponent>;
 
   public items: any[] = [
     { name: 'Option 1', disabled: false },
@@ -46,41 +100,33 @@ export class DropdownTestComponent {
     { name: 'Option 4', disabled: false }
   ];
 
-  @ViewChild('dropdown', {
-    read: SkyDropdownComponent,
-    static: false
-  })
-  public dropdown: SkyDropdownComponent;
-
-  @ViewChild('dropdownMenu', {
-    read: SkyDropdownMenuComponent,
-    static: false
-  })
-  public dropdownMenu: SkyDropdownMenuComponent;
-
-  @ViewChild('outsideButton', {
-    read: ElementRef,
-    static: false
-  })
-  public outsideButton: ElementRef;
-
   constructor(
     private changeDetector: ChangeDetectorRef
   ) { }
 
-  public sendMessage(type: SkyDropdownMessageType) {
-    this.dropdownController.next({ type });
-  }
+  public onMenuChanges(): void { }
 
   public changeItems() {
     this.items.pop();
     this.changeDetector.detectChanges();
   }
 
-  public setItems(items: any[]) {
+  public setManyItems(): void {
+    const items: any[] = [];
+
+    for (let i = 0; i < 50; i++) {
+      items.push({
+        name: `Option ${i}`,
+        disabled: false
+      });
+    }
+
     this.items = items;
-    this.changeDetector.detectChanges();
+    this.changeDetector.markForCheck();
   }
 
-  public onDropdownClick() {}
+  public sendMessage(type: SkyDropdownMessageType) {
+    this.messageStream.next({ type });
+  }
+
 }