+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 3.5.1 (2020-04-08)
+
+- Added a deprecation message to the `SkyPopoverTrigger` type. [#63](https://github.com/blackbaud/skyux-popovers/pull/63)
+- Fixed the dropdown component to properly set all ARIA attributes. [#67](https://github.com/blackbaud/skyux-popovers/pull/67)
+- Fixed `SkyPopoverContext` to stop throwing build warnings. [#63](https://github.com/blackbaud/skyux-popovers/pull/63)
+
+# 3.5.0 (2020-04-02)
+
+- Updated the dropdown and popover components to implement the affix and overlay services. [#61](https://github.com/blackbaud/skyux-popovers/pull/61)
+
 # 4.0.0-rc.0 (2020-02-19)
 
 ### New features